Update type inference in parameters.py

+++ ema_workbench/em_framework/parameters.py
@@ -604,5 +604,5 @@ def create_parameters(uncertainties, **kwargs):
         if type == 'cat':
             uncs.append(parameter_map[type](name, values))
         else:
-            uncs.append(parameter_map[type](name, *values))
+            uncs.append(parameter_map[type](name, *values.astype(float)))
     return uncs
